Add float support to the input of the MLP

+++ micrograd/build_graph.py
@@ -11,7 +11,7 @@ def main():
         "--save_dir", "-s", type=str, help="Directory to save the Computational Graph"
     )
     parser.add_argument(
-        "--input", "-i", required=True, nargs="+", type=int, help="Input of the MLP"
+        "--input", "-i", required=True, nargs="+", type=float, help="Input of the MLP"
     )
     parser.add_argument(
         "--output", "-o", required=True, nargs="+", type=int, help="A list of neurons output"
